Fixes all around -- now some cool demos work (#395)

+++ impl/src/tooling/icpp/interpreter/runner.cpp
@@ -413,7 +413,7 @@ int main(int argc, char** argv)
         {
             printf("!ERROR!\n");
         }
-        printf("Elapsed time %lli...\n", (int64_t)delta_ms);
+        printf("Elapsed time %i...\n", (int)delta_ms);
 
         return 0;
     }
